+++ hw09_generator_of_validators/go-validate/main.go
@@ -16,7 +16,5 @@ func writeToFile(str string, path string) {
 
 func main() {
 	res := generateMultipleStructValidations(parseAST())
-
 	writeToFile(res, "models/models_validation_generated.go")
-
 }
